+++ src/com/DirectionsValues.java
@@ -1,3 +1,15 @@
+/**
+ * @author Created by Haubir -  haubir.mariwani@fasbros.it
+ *                              jagheterhaubir@gmail.com
+ *
+ * The important values of a travel directions description.
+ * <p>
+ * The DirectionsValues object contains detailed information from the description for travel directions, retrieved from
+ * the Google Maps Directions API.
+ *
+ * @version //What version of java is current
+ * @since April 11th of 2017
+ */
 package com;
 
 import com.google.maps.DirectionsApi;
@@ -12,13 +24,10 @@ import static com.google.maps.model.TravelMode.DRIVING;
 import static com.google.maps.model.TravelMode.TRANSIT;
 import static com.google.maps.model.TravelMode.WALKING;
 
-/**
- * Created by Haubir on 4/7/17.
- */
 public class DirectionsValues {
     private DirectionsResult dirResult;
-    private int totalDistance;
-    private int totalDuration;
+    private long totalDistance;
+    private long totalDuration;
 /*
     public DirectionsValues(GeoApiContext context, String origAddress, String destAddress) throws InterruptedException, ApiException, IOException {
         new DirectionsValues(context, origAddress, destAddress, DRIVING);
@@ -28,8 +37,10 @@ public class DirectionsValues {
 */
     public DirectionsValues(GeoApiContext context, String origAddress, String destAddress, TravelMode mode) throws InterruptedException, ApiException, IOException {
         this.dirResult = DirectionsApi.getDirections(context, origAddress, destAddress).mode(mode).await();
-        this.totalDistance = (int) this.dirResult.routes[0].legs[0].distance.inMeters;
-        this.totalDuration = (int) this.dirResult.routes[0].legs[0].duration.inSeconds;
+        //this.totalDistance = (int) this.dirResult.routes[0].legs[0].distance.inMeters;
+        //this.totalDuration = (int) this.dirResult.routes[0].legs[0].duration.inSeconds;
+        this.totalDistance = this.dirResult.routes[0].legs[0].distance.inMeters;
+        this.totalDuration = this.dirResult.routes[0].legs[0].duration.inSeconds;
         System.out.println("totalDistance: " + this.totalDistance);
         System.out.println("totalDuration: " + this.totalDuration);
     }
@@ -38,11 +49,11 @@ public class DirectionsValues {
         return dirResult;
     }
 
-    public int getTotalDistance() {
+    public long getTotalDistance() {
         return totalDistance;
     }
 
-    public int getTotalDuration() {
+    public long getTotalDuration() {
         return totalDuration;
     }
 }
